style(backend): unified interface for executing commands
