+++ src-tauri/src/launch/commands.rs
@@ -1,5 +1,5 @@
 use super::helpers::command_generator::{
-  export_full_launch_command, generate_launch_command, split_wrapper, LaunchCommand,
+  export_full_launch_command, generate_launch_command, LaunchCommand,
 };
 use super::helpers::file_validator::{
   extract_native_libraries, get_invalid_assets, get_invalid_library_files,
@@ -23,6 +23,7 @@ use crate::resource::helpers::misc::get_source_priority_list;
 use crate::storage::load_json_async;
 use crate::tasks::commands::schedule_progressive_task_group;
 use crate::utils::fs::create_zip_from_dirs;
+use crate::utils::shell::{execute_command_line, split_wrapper};
 use crate::utils::window::create_webview_window;
 use std::collections::HashMap;
 use std::fs;
@@ -270,20 +271,9 @@ pub async fn launch_game(
   let full_cmd = export_full_launch_command(&class_paths, &cmd_args, &selected_java.exec_path);
   println!("[Launch Command] {}", full_cmd);
 
-  let precall = game_config.advanced.custom_commands.precall_command.clone();
-  if !precall.trim().is_empty() {
-    #[cfg(target_os = "windows")]
-    let _ = std::process::Command::new("cmd")
-      .arg("/C")
-      .arg(&precall)
-      .creation_flags(0x08000000)
-      .status();
-
-    #[cfg(not(target_os = "windows"))]
-    let _ = std::process::Command::new("/bin/sh")
-      .arg("-c")
-      .arg(&precall)
-      .status();
+  let precall_cmd = game_config.advanced.custom_commands.precall_command.clone();
+  if !precall_cmd.trim().is_empty() {
+    let _ = execute_command_line(&precall_cmd);
   }
 
   // execute launch command
@@ -308,7 +298,7 @@ pub async fn launch_game(
     launching.full_command = full_cmd;
   }
 
-  let post_exit = game_config
+  let post_exit_cmd = game_config
     .advanced
     .custom_commands
     .post_exit_command
@@ -325,7 +315,7 @@ pub async fn launch_game(
     &game_config.game_window.custom_title,
     game_config.launcher_visibility.clone(),
     tx,
-    Some(post_exit),
+    Some(post_exit_cmd),
   )
   .await?;
   let _ = rx.recv();
