+++ src-tauri/src/utils/mod.rs
@@ -2,6 +2,7 @@ pub mod commands;
 pub mod fs;
 pub mod image;
 pub mod portable;
+pub mod shell;
 pub mod string;
 pub mod sys_info;
 pub mod web;
